[SPARK-KERNELS]: Fix parallel spark cluster up. Fix reverse-proxy links and patterns.

+++ infrastructure-provisioning/src/general/scripts/os/tensor-rstudio_install_dataengine_kernels.py
@@ -40,20 +40,20 @@ args = parser.parse_args()
 def configure_notebook(keyfile, hoststring):
     scripts_dir = '/root/scripts/'
     templates_dir = '/root/templates/'
-    put(scripts_dir + 'tensor-rstudio_dataengine_create_configs.py', '/tmp/tensor-rstudio_dataengine_create_configs.py')
-    put(templates_dir + 'notebook_spark-defaults_local.conf', '/tmp/notebook_spark-defaults_local.conf')
+    run('mkdir -p /tmp/{}/'.format(args.cluster_name))
+    put(templates_dir + 'notebook_spark-defaults_local.conf', '/tmp/{}/notebook_spark-defaults_local.conf'.format(args.cluster_name))
     spark_master_ip = args.spark_master.split('//')[1].split(':')[0]
     spark_memory = get_spark_memory(True, args.os_user, spark_master_ip, keyfile)
-    run('echo "spark.executor.memory {0}m" >> /tmp/notebook_spark-defaults_local.conf'.format(spark_memory))
-    sudo('\cp /tmp/tensor-rstudio_dataengine_create_configs.py /usr/local/bin/tensor-rstudio_dataengine_create_configs.py')
-    sudo('chmod 755 /usr/local/bin/tensor-rstudio_dataengine_create_configs.py')
-    sudo('mkdir -p /usr/lib/python2.7/dlab/')
-    run('mkdir -p /tmp/dlab_libs/')
-    local('scp -i {} /usr/lib/python2.7/dlab/* {}:/tmp/dlab_libs/'.format(keyfile, hoststring))
-    run('chmod a+x /tmp/dlab_libs/*')
-    sudo('mv /tmp/dlab_libs/* /usr/lib/python2.7/dlab/')
-    if exists('/usr/lib64'):
-        sudo('ln -fs /usr/lib/python2.7/dlab /usr/lib64/python2.7/dlab')
+    run('echo "spark.executor.memory {0}m" >> /tmp/{1}/notebook_spark-defaults_local.conf'.format(spark_memory, args.cluster_name))
+    if not exists('/usr/local/bin/tensor-rstudio_dataengine_create_configs.py'):
+        put(scripts_dir + 'tensor-rstudio_dataengine_create_configs.py', '/usr/local/bin/tensor-rstudio_dataengine_create_configs.py', use_sudo=True)
+        sudo('chmod 755 /usr/local/bin/tensor-rstudio_dataengine_create_configs.py')
+    if not exists('/usr/lib/python2.7/dlab/'):
+        sudo('mkdir -p /usr/lib/python2.7/dlab/')
+        put('/usr/lib/python2.7/dlab/*', '/usr/lib/python2.7/dlab/', use_sudo=True)
+        sudo('chmod a+x /usr/lib/python2.7/dlab/*')
+        if exists('/usr/lib64'):
+            sudo('ln -fs /usr/lib/python2.7/dlab /usr/lib64/python2.7/dlab')
 
 
 if __name__ == "__main__":
